+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- [Enhancement] Register as platform for dynamically adding services after connecting ([#855](https://github.com/homebridge-xiaomi-roborock-vacuum/homebridge-xiaomi-roborock-vacuum/pull/855))
+
 ## 0.34.0
 
 - [New Model] Add waterbox support for Roborock S7 Max Ultra. ([#1027](https://github.com/homebridge-xiaomi-roborock-vacuum/homebridge-xiaomi-roborock-vacuum/pull/1027))